Optimize update_user_groups to a fixed number of DB queries.
By using Django's bulk_create and set operations we can reduce the
time complexity of this function from two sorts to None and from
3 + (2ish * (claim_groups - existing _group_names) OR (
claim_groups - existing _group_names) ) DB queries to a constant 3.

The downside of this approach is that the bulk_create operation
will not cause the pre_save and post_save signals to be raised
when creating the Group instances.