Check the new groups are different before update
Claimed groups can contain groups they are not relevant for our app
so the first quick check is to ensure the list are different then
to use a SQL query to fetch the groups of the user.
Only set the groups if they are different, it avoids a SELECT of
the Django ORM (called by the set()).